\name{Flowmatrix}
\alias{Flowmatrix}
\title{
  Generates a flow matrix for an inverse (foodweb) problem
}
\description{
  Given a linear inverse model food web input list, generates a flow
  matrix, that contains the values of flows
}

\usage{
Flowmatrix(lim, web = NULL)
}

\arguments{
  \item{lim }{a list that contains the linear inverse model
    specification, as  generated by function \code{setup.limfile}.
  }
  \item{web }{the solved (food) web problem, i.e. the values of the
    unknowns; if not specified, the model is solved first, using \code{Lsei}
  }
}
\value{
   the flow matrix, containing the magnitude of the flows.

   The value on row i and column j is the flow *from* i and *to* j
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\examples{
Flowmatrix(LIMRigaAutumn)
}
\seealso{
  \code{\link[diagram]{plotweb}} from package \code{diagram},
  which takes as input the flow matrix and plots the food web
}

\keyword{utilities}