\name{Variables}
\alias{Variables}
\title{
  Generates the values of variables for a linear inverse (foodweb) problem
}

\description{
  Given an linear inverse model input list, generates the values of the
  inverse variables
}

\usage{
Variables (lim, res, ...)
}
\arguments{
  \item{lim }{a list that contains the linear inverse model
    specification, as  generated by function \code{setup.limfile}.
  }
  \item{res }{the solved linear inverse problem; if not specified,
    the model is solved first, using \code{Lsei.lim<}.
  }
  \item{... }{extra parameters passed to function \code{Lsei.lim}.
  }
}

\value{
  the variables, a one-column data.frame
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\examples{
 Variables(LIMRigaAutumn)
}

\seealso{
  \code{\link{Varranges}} which estimates the ranges of variables.
}
\keyword{utilities}

