\name{Varranges}
\alias{Varranges}
\title{
  Generates ranges of the variables for a linear inverse problem
}
\description{
  Given an inverse input list, generates the minimal and maximal values of
  the variables (linear combinations of unknowns).
}
\usage{
Varranges(lim, ...)
}
\arguments{
  \item{lim }{a list that contains the linear inverse model
    specification, as  generated by function \code{setup.limfile}.
  }
  \item{... }{extra arguments passed to function \code{varranges}.
  }
}
\value{
  a 2-columned vector containing the minimum (column 1) and
  maximum (column 2) of each variable.
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}

\examples{
Varranges(LIMRigaAutumn)
}

\seealso{
  \code{\link{Xranges}} which estimates the ranges of unknowns

  \code{\link{Plotranges}} to plot the ranges
}

\keyword{optimize}
\keyword{utilities}

