\name{Plotranges}
\alias{Plotranges}
\alias{Plotranges.double}
\alias{Plotranges.lim}
\alias{Plotranges.character}
\title{Plots minimum and maximum ranges.
Takes as input either a lim list, as generated by Setup or a set of vectors specifying
the minimu, maximum and the central value, or a data.frame that contains min, max and central values.
}
\description{Plots the minimum and maximum and central values}
\usage{
Plotranges.double(min, max, value=NULL, labels=NULL,log="",pch=16,
pch.col="black", line.col="gray", seg.col="black", xlim=NULL, 
main=NULL, xlab=NULL, ylab=NULL, lab.cex=1.0, ...)
Plotranges.lim(lim=NULL,labels=NULL,log="",pch=16,
pch.col="black", line.col="gray", seg.col="black", xlim=NULL,
main=NULL, xlab=NULL, ylab=NULL, lab.cex=1.0, ...)
Plotranges.character(file,...)
Plotranges(...)
}
\arguments{
  \item{min }{minimum value}
  \item{max }{maximum value}
  \item{value }{median or mean value}
  \item{lim }{a list that contains the linear inverse model specification, as  generated by function \code{Setup}}
  \item{file }{name of the linear inverse input file}
  \item{labels }{names of each value}
  \item{log }{if = x: logarithmic scale for x-axis}
  \item{pch }{pch symbol used for mean value}
  \item{pch.col }{pch color for mean value}
  \item{line.col }{color for each variable, spanning x-axis}
  \item{seg.col }{color for variable range}
  \item{xlim }{limits on x-axis}
  \item{main }{main title}
  \item{xlab }{x-axis label}
  \item{ylab }{y-axis label}
  \item{lab.cex }{label expansion value}
  \item{... }{arguments passed to R-function "text" when writing labels}
}
\author{Karline Soetaert <k.soetaert@nioo.knaw.nl>}
\value{
  Only when a lim list was inputted. A data frame with
  \item{min }{the minimum }
  \item{max }{the maximum }
  \item{values }{the central value }
}
\keyword{hplot}

