\name{PrintMat}
\alias{PrintMat}
\title{
  Writes linear inverse matrices to the screen
}

\description{
  Prints the linear inverse problem:

  \itemize{
    \item inverse matrices and vectors \code{A}, \code{b} of the
      equalities \deqn{Ax=b}
    \item inverse matrices and vectors \code{G}, \code{h} of the
      inequalities \deqn{Gx>=h}
    \item if present, also writes the cost/profit function
  }
}

\usage{
PrintMat(lim)
}

\arguments{
  \item{lim }{a list that contains the linear inverse model
    specification, as  generated by function \code{setup.limfile}.
  }
}
\value{
returns nothing.
}
\author{Karline Soetaert <karline.soetaert@nioz.nl>}
\examples{
   PrintMat(LIMBlending)
}

\keyword{IO}
\keyword{utilities}

